We now have basic progress updates working on both Jupyter and command line, for the single-threaded sequential executor.  Next we need to do it for the parallel and pipelined executors
